requirements.txt
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
langchain-community==0.3.13
python-dotenv==1.0.0
openai==1.59.3
faiss-cpu==1.9.0
tiktoken==0.8.0